pyproject.toml
[tool.poetry]
name = "test"
version = "0.1.0"
description = ""
authors = ["AndresJalife <ajalife@fi.uba.ar>"]

[tool.poetry.dependencies]
python = "^3.8"
SQLAlchemy = "^1.4.35"
requests = "^2.27.1"
click = "^8.1.2"
psycopg2-binary = "^2.9.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
